moved LoggedIn state to loggenIn folder
